+++ global-client/src/logic/store/settings.ts
@@ -1,2 +1,2 @@
-import { Theme, TimeFormat, defaultSettings, FeatureFlag, useSettings, i18nSettings } from '../../../../hub-client/src/logic/store/settings';
-export { Theme, TimeFormat, defaultSettings, FeatureFlag, useSettings, type i18nSettings };
+import { Theme, TimeFormat, NotificationsPermission, defaultSettings, FeatureFlag, useSettings, i18nSettings } from '../../../../hub-client/src/logic/store/settings';
+export { Theme, TimeFormat, NotificationsPermission, defaultSettings, FeatureFlag, useSettings, type i18nSettings };